% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_iata.R
\name{import_iata}
\alias{import_iata}
\title{Import IATA data}
\usage{
import_iata(filename)
}
\arguments{
\item{filename}{The file path that should be imported}
}
\value{
A dataframe.
}
\description{
Import IATA data
}
