% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_iata_multi.R
\name{import_iata_multi_nat}
\alias{import_iata_multi_nat}
\title{Import multiple IATA datasets at a national level}
\usage{
import_iata_multi_nat(analysis_years, folder)
}
\arguments{
\item{analysis_years}{The years for which IATA data should be imported.}

\item{folder}{The file path for the folder where IATA data are saved.}
}
\value{
A dataframe formed from the flight passenger numbers in all the
imported datasets. Flight passenger numbers are aggregated to a national
level in which the origin and destination locations are countries (rather
than individual airports as in the raw IATA data).
}
\description{
Import multiple IATA datasets at a national level
}
\examples{
analysis_years <- 2012:2014
folder <- "~/iata_analysis/data/"
import_iata_multi_nat(analysis_years, folder)
}
