% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_matrix.R
\name{od_matrix}
\alias{od_matrix}
\title{Create O-D matrix from flight passenger numbers}
\usage{
od_matrix(flight_data, flight_month, flight_year)
}
\arguments{
\item{flight_data}{The dataframe containing the flight passenger numbers.}

\item{month}{The month (or months) that the flight data used to construct
the O-D matrix should cover.}

\item{year}{The year (or years) that the flight data used to construct
the O-D matrix should cover.}
}
\value{
Matrix in which each cell represents the probability that a person
leaving the origing country (the rows of the matrix) flies to a given destination
(the columns of the matrix).
}
\description{
Create O-D matrix from flight passenger numbers
}
